refactoring code from fetch.js into utils file as handleUserRequest() function
correcting selection error
