+++ Downloads/book_app_8th/database.json
@@ -1,1 +1,1 @@
-{"users":{"ephie":{"readingList":["Book ID :0 {\n              title:Valhalla's Swordsmith,\n              author:Tom Hill,\n              publishingCompany:Andrews UK Limited,\n         }"]},"nate":{"readingList":["{\n              title:Scythe,\n              author:Neal Shusterman,\n              publishingCompany:Simon and Schuster,\n         }"]},"john                     ":{"readingList":["{\n              title:The Fault in Our Stars,\n              author:John Green,\n              publishingCompany:Penguin UK,\n         }"]},"brina":{"readingList":["{\n      title:Anne of Manhattan,\n      author:Brina Starler,\n      publishingCompany:HarperCollins,\n }"]},"jenkins":{"readingList":["{\n      title:The Seven Husbands of Evelyn Hugo,\n      author:Taylor Jenkins Reid,\n      publishingCompany:Simon and Schuster,\n }"]},"tom":{"readingList":["{\n      title:Basher Science: An A to Z of Science,\n      author:Tom Jackson,\n      publishingCompany:Kingfisher,\n }"]},"bolu":{"readingList":["{\n      title:Love in Colour,\n      author:Bolu Babalola,\n      publishingCompany:Publisher not found,\n }"]},"james":{"readingList":["{\n      title:Vue.js: 11 Practical Projects,\n      author:James Hibbard,Yomi Eluwande,Michael Wanyoike,Nilson Jacques,Christopher Vundi,Deji Atoyebi,Ivaylo Gerchev,\n      publishingCompany:SitePoint,\n }"]},"brian":{"readingList":["{\n      title:Beauty Will Save the World,\n      author:Brian Zahnd,\n      publishingCompany:Charisma Media,\n }","{\n      title:Saving Container Plants,\n      author:Brian McGowan,Alice McGowan,\n      publishingCompany:Storey Publishing,\n }"]}}}
+{"users":{"ken":{"readingList":["{\n      title:Tokyo Revengers 10,\n      author:Ken Wakui,\n      publishingCompany:Kodansha America LLC,\n }","{\n      title:Tokyo Revengers 1,\n      author:Ken Wakui,\n      publishingCompany:Kodansha America LLC,\n }","{\n      title:Tokyo Revengers 4,\n      author:Ken Wakui,\n      publishingCompany:Kodansha America LLC,\n }","{\n      title:Tokyo Revengers 6,\n      author:Ken Wakui,\n      publishingCompany:Kodansha America LLC,\n }"]}}}
